BOJ13305 정답 추가

+++ src/BOJ-13305.h
@@ -7,6 +7,6 @@ using namespace std;
 
 namespace BOJ13305 {
 	unsigned long long solution(int numOfCity, vector<int> oilPriceEachCity, vector<int> distanceEachCity);
-	int getTotalDistance(int curCityIdx, int disCityIdx, vector<int> distanceEachCity);
+	unsigned long long getTotalDistance(int curCityIdx, int disCityIdx, vector<int> distanceEachCity);
 	void bojMain();
 }
